backend/requirements-backend.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6